requirements.txt
azure-functions
psutil 